+++ packages/gamut/src/Alert/__tests__/Alert-test.tsx
@@ -98,6 +98,20 @@ describe('Alert', () => {
     expect(renderedTruncated.prop('lines')).toEqual(2);
   });
 
+  it('automatically displays content as expanded if no lines are set', () => {
+    const renderedAlert = mount(
+      <Alert
+        onClose={onClose}
+        cta={{ text: 'Click Me', onClick: onCtaClick, href: '/hello' }}
+      >
+        Hello
+      </Alert>
+    );
+
+    const renderedTruncated = renderedAlert.find('Truncate');
+    expect(renderedTruncated.prop('expanded')).toEqual(true);
+  });
+
   it('renders a clickable button to expand the truncated section', () => {
     const renderedAlert = mount(
       <Alert
@@ -116,7 +130,7 @@ describe('Alert', () => {
 
     renderedAlert.update();
 
-    expect(renderedAlert.find('Truncate').prop('lines')).toEqual(undefined);
+    expect(renderedAlert.find('Truncate').prop('expanded')).toEqual(true);
   });
 
   it('renders a clickable button to collapse the truncated section when clicked twice', () => {
@@ -131,6 +145,7 @@ describe('Alert', () => {
     );
 
     const buttons = renderedAlert.find('button');
+    expect(buttons.length).toBe(3);
 
     buttons.at(0).simulate('click');