+++ packages/gamut/src/Alert/index.tsx
@@ -77,7 +77,8 @@ export const Alert: React.FC<AlertProps> = ({
             justify="spaceBetween"
           >
             <Truncate
-              lines={isExpanded ? undefined : lines}
+              expanded={isExpanded || !Boolean(lines)}
+              lines={lines || 1}
               onTruncate={setIsTruncated}
               className={styles.truncate}
             >