+++ packages/gamut/src/Overlay/index.tsx
@@ -1,5 +1,6 @@
 import FocusTrap from 'focus-trap-react';
-import React, { useLayoutEffect } from 'react';
+import React from 'react';
+import { useIsomorphicLayoutEffect } from 'react-use';
 
 import { BodyPortal } from '../BodyPortal';
 import { FlexBox } from '../Box';
@@ -40,7 +41,7 @@ export const Overlay: React.FC<OverlayProps> = ({
   onRequestClose,
   isOpen,
 }) => {
-  useLayoutEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     document.body.style.overflow = isOpen ? 'hidden' : 'visible';
   }, [isOpen]);